Influence bar in DiplomacyScreen
